+++ cypress/integration/docs-smoke-test/top-menu.js
@@ -2,7 +2,10 @@ import { URL_DOCS_SMOKE_TEST } from '../../support/urls';
 
 describe('Top menu', () => {
   it('should toggle top menu and take a snapshot', () => {
-    cy.visit(URL_DOCS_SMOKE_TEST);
+    // Wait for Gastby to be fully loaded otherwise the click event won't be
+    // handled correctly.
+    // We can check for the existence of the `data-react-helmet` in the `html` document.
+    cy.visit(URL_DOCS_SMOKE_TEST).get('html[data-react-helmet]');
     cy.findByLabelText('Open Top Menu').click();
     cy.findByRole('top-menu').should('be.visible');
     cy.findByRole('top-menu').within(() => {
@@ -12,7 +15,10 @@ describe('Top menu', () => {
     cy.findByLabelText('Close Top Menu').should('exist');
   });
   it('should close top menu when clicking on the search input', () => {
-    cy.visit(URL_DOCS_SMOKE_TEST);
+    // Wait for Gastby to be fully loaded otherwise the click event won't be
+    // handled correctly.
+    // We can check for the existence of the `data-react-helmet` in the `html` document.
+    cy.visit(URL_DOCS_SMOKE_TEST).get('html[data-react-helmet]');
     cy.findByLabelText('Open Top Menu').click();
     cy.findByRole('top-menu').should('be.visible');
     cy.findByRole('top-menu').within(() => {