+++ packages/gatsby-theme-api-docs/src/components/type/properties/rows/name-type.js
@@ -1,11 +1,11 @@
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { css } from '@emotion/react';
 import { designSystem, Markdown } from '@commercetools-docs/ui-kit';
 import SpacingsStack from '@commercetools-uikit/spacings-stack';
 import SpacingsInline from '@commercetools-uikit/spacings-inline';
 import { BetaFlag } from '@commercetools-docs/gatsby-theme-docs';
 import { typography } from '../../../../design-system';
+import RegexProperty from '../../properties/regex-properties';
 import useTypeToRender from '../../../../hooks/use-type-to-render';
 import Required from '../../../required';
 
@@ -14,8 +14,12 @@ import Required from '../../../required';
 // into lines before the name is wrapped in itself if it consists of multiple words.
 const PropertyName = styled.div`
   white-space: nowrap;
+  line-height: ${typography.lineHeights.propertyType};
 `;
-const PropertyType = styled.div``;
+const PropertyType = styled.div`
+  line-height: ${typography.lineHeights.propertyType};
+  color: ${designSystem.colors.light.textFaded};
+`;
 const BetaWrapper = styled.span`
   font-size: ${typography.fontSizes.body};
 `;
@@ -29,35 +33,16 @@ const NameType = (props) => {
   const isRegex = (string) =>
     string.charAt(0) === '/' && string.charAt(string.length - 1) === '/';
 
-  const getExpressionInsideSlashes = (input) => {
-    return input.match(/\/([^;]*)\//);
-  };
-
   return (
     <SpacingsStack scale="xs">
-      <PropertyName className="name-type">
+      <PropertyName>
         <SpacingsInline scale="xs">
           {isRegex(props.property.name) ? (
-            <Markdown.InlineCode>
-              <span
-                css={css`
-                  color: ${designSystem.colors.light.textInfo};
-                `}
-              >
-                /
-              </span>
-              {getExpressionInsideSlashes(props.property.name)[1]}
-              <span
-                css={css`
-                  color: ${designSystem.colors.light.textInfo};
-                `}
-              >
-                /
-              </span>
-            </Markdown.InlineCode>
+            <RegexProperty expression={props.property.name} />
           ) : (
             <Markdown.InlineCode>{props.property.name}</Markdown.InlineCode>
           )}
+          {'\u200B' /* zero-width space for the search crawler */}
           {props.property.required && <Required />}
           {props.property.beta && (
             <BetaWrapper>
@@ -66,19 +51,15 @@ const NameType = (props) => {
           )}
         </SpacingsInline>
       </PropertyName>
-      <PropertyType className="name-type">
-        {typeToRender.displayPrefix && (
-          <span className="name">{typeToRender.displayPrefix}</span>
-        )}
-        {isRegex(props.property.name) ? (
-          <span className="name">
-            Any property matching regular expression above
-          </span>
-        ) : typeof typeToRender.type === 'string' ? (
-          <span className="name">{typeToRender.type}</span>
-        ) : (
-          typeToRender.type
-        )}
+      <PropertyType>
+        {typeToRender.displayPrefix && typeToRender.displayPrefix}
+
+        {isRegex(props.property.name)
+          ? `Any ${typeToRender.type.toLowerCase()} property matching this regular expression`
+          : typeof typeToRender.type === 'string'
+          ? typeToRender.type
+          : typeToRender.type}
+        {'\u200B' /* zero-width space for the search crawler */}
       </PropertyType>
     </SpacingsStack>
   );