+++ packages/gatsby-theme-docs/gatsby-node.js
@@ -43,11 +43,11 @@ exports.onPreBootstrap = (gatsbyApi, themeOptions) => {
 exports.createResolvers = ({ createResolvers }) => {
   const resolvers = {
     SiteSiteMetadata: {
-      // this field is needed by 'gatsby-plugin-feed' plugin
+      // this field is needed by plugins needing an absolute production site URL, e.g. the'gatsby-plugin-feed' plugin
       siteUrl: {
         type: 'String',
-        resolve(source, args, context) {
-          const site = context.nodeModel.getAllNodes({ type: 'Site' })[0];
+        resolve: async (source, args, context) => {
+          const site = await context.nodeModel.findOne({ type: 'Site' });
           return `https://${source.productionHostname}${site.pathPrefix}`;
         },
       },
@@ -483,11 +483,18 @@ exports.onCreateWebpackConfig = (
         {
           loader: require.resolve('@svgr/webpack'),
           options: {
-            // NOTE: disable this and manually add `removeViewBox: false` in the SVGO plugins list
-            // See related PR: https://github.com/smooth-code/svgr/pull/137
             icon: false,
             svgoConfig: {
-              plugins: [{ removeViewBox: false }, { cleanupIDs: true }],
+              plugins: [
+                {
+                  name: 'preset-default',
+                  params: {
+                    overrides: {
+                      removeViewBox: false,
+                    },
+                  },
+                },
+              ],
             },
           },
         },