+++ packages/gatsby-transformer-raml/src/schema/define-raml-resource.js
@@ -20,7 +20,9 @@ const defineRamlResource = ({ schema, createTypes }) => {
         uriParameters: '[RamlResourceUriParameter!]',
         post: 'RamlResourceMethodWithBody',
         put: 'RamlResourceMethodWithBody',
+        patch: 'RamlResourceMethodWithBody',
         get: 'RamlResourceMethodWithoutBody',
+        head: 'RamlResourceMethodWithoutBody',
         delete: 'RamlResourceMethodWithoutBody',
       },
       interfaces: ['Node'],
@@ -98,7 +100,7 @@ const defineRamlResource = ({ schema, createTypes }) => {
       fields: {
         code: 'Int!',
         description: 'String',
-        body: 'RamlResourceMethodBody!',
+        body: 'RamlResourceMethodBody',
       },
     }),