+++ websites/site-template/gatsby-config.js
@@ -6,7 +6,6 @@ const {
 
 module.exports = {
   flags: {
-    DEV_SSR: true,
     FAST_DEV: true,
   },
   pathPrefix: '/site-template',